# Support bracket notation for jsonpath