--- a/src/main/java/org/embulk/filter/column/JsonColumn.java
+++ b/src/main/java/org/embulk/filter/column/JsonColumn.java
@@ -1,6 +1,11 @@
 package org.embulk.filter.column;
 
 import org.embulk.config.ConfigException;
+import org.embulk.filter.column.path.ArrayPathToken;
+import org.embulk.filter.column.path.CompiledPath;
+import org.embulk.filter.column.path.PathCompiler;
+import org.embulk.filter.column.path.PathToken;
+import org.embulk.filter.column.path.PropertyPathToken;
 import org.embulk.spi.type.Type;
 import org.msgpack.value.StringValue;
 import org.msgpack.value.Value;
@@ -39,31 +44,43 @@ public class JsonColumn
 
     public JsonColumn(String path, Type type, Value defaultValue, String src)
     {
-        this.path = path;
+        CompiledPath compiledPath = PathCompiler.compile(path);
+        CompiledPath compiledSrc = src == null ? compiledPath : PathCompiler.compile(src);
+        this.path = compiledPath.toString();
         this.type = type;
         this.defaultValue = (defaultValue == null ? ValueFactory.newNil() : defaultValue);
-        this.src = (src == null ? path : src);
+        this.src = compiledSrc.toString();
 
         this.pathValue = ValueFactory.newString(path);
-        this.parentPath = parentPath(path);
-        this.baseName = baseName(path);
+        this.parentPath = compiledPath.getParentPath();
+        this.baseName = compiledPath.getTailPath();
         if (this.baseName.equals("[*]")) {
             throw new ConfigException(String.format("%s wrongly ends with [*], perhaps you can remove the [*]", path));
         }
-        this.baseIndex = baseIndex(path);
+        this.baseIndex = compiledPath.baseIndex();
         this.parentPathValue = ValueFactory.newString(parentPath);
+        String baseName = getPropertyOrBaseName(compiledPath, this.baseName);
         this.baseNameValue = ValueFactory.newString(baseName);
 
         this.srcValue = ValueFactory.newString(this.src);
-        this.srcParentPath = parentPath(this.src);
-        this.srcBaseName = baseName(this.src);
-        this.srcBaseIndex = baseIndex(this.src);
+        this.srcParentPath = compiledSrc.getParentPath();
+        this.srcBaseName = compiledSrc.getTailPath();
+        this.srcBaseIndex = compiledSrc.baseIndex();
         this.srcParentPathValue = ValueFactory.newString(this.srcParentPath);
-        this.srcBaseNameValue = ValueFactory.newString(this.srcBaseName);
+        String srcBaseName = getPropertyOrBaseName(compiledSrc, this.srcBaseName);
+        this.srcBaseNameValue = ValueFactory.newString(srcBaseName);
 
         if (! srcParentPath.equals(parentPath)) {
             throw new ConfigException(String.format("The branch (parent path) of src \"%s\" must be same with of name \"%s\" yet", src, path));
         }
+    }
+
+    private String getPropertyOrBaseName(CompiledPath path, String baseName) {
+      if (path.getTail() instanceof PropertyPathToken) {
+          return ((PropertyPathToken) path.getTail()).getProperty();
+      } else {
+          return baseName;
+      }
     }
 
     public String getPath()
@@ -149,46 +166,20 @@ public class JsonColumn
     // like File.dirname
     public static String parentPath(String path)
     {
-        String[] parts = path.split("\\.");
-        StringBuilder builder = new StringBuilder();
-        for (int i = 0; i < parts.length - 1; i++) {
-            builder.append(".").append(parts[i]);
-        }
-        if (parts[parts.length - 1].contains("[")) {
-            String[] arrayParts = parts[parts.length - 1].split("\\[");
-            builder.append(".").append(arrayParts[0]);
-            for (int j = 1; j < arrayParts.length - 1; j++) {
-                builder.append("[").append(arrayParts[j]);
-            }
-        }
-        return builder.deleteCharAt(0).toString();
+        return PathCompiler.compile(path).getParentPath();
     }
 
     public static String baseName(String path)
     {
-        String[] parts = path.split("\\.");
-        String[] arrayParts = parts[parts.length - 1].split("\\[");
-        if (arrayParts.length == 1) { // no [i]
-            return arrayParts[arrayParts.length - 1];
-        }
-        else {
-            return "[" + arrayParts[arrayParts.length - 1];
-        }
+        return PathCompiler.compile(path).getTailPath();
     }
 
     public static Long baseIndex(String path)
     {
-        String baseName = baseName(path);
-        if (baseName.startsWith("[") && baseName.endsWith("]")) {
-            String baseIndex = baseName.substring(1, baseName.length() - 1);
-            try {
-                return Long.parseLong(baseIndex);
-            }
-            catch (NumberFormatException e) {
-                return null;
-            }
-        }
-        else {
+        PathToken pathToken = PathCompiler.compile(path).getTail();
+        if (pathToken instanceof ArrayPathToken) {
+            return ((ArrayPathToken) pathToken).index().longValue();
+        } else {
             return null;
         }
     }