--- a/src/main/java/org/embulk/filter/column/path/PathTokenAppender.java
+++ b/src/main/java/org/embulk/filter/column/path/PathTokenAppender.java
@@ -1,0 +1,6 @@
+package org.embulk.filter.column.path;
+
+public interface PathTokenAppender
+{
+    PathTokenAppender appendPathToken(PathToken next);
+}