Support the bracket notation using the [JsonPath](https://github.com/jayway/JsonPath) partially. (PathCompiler related)
The porting parts from JsonPath have been some changes from the original and located under the the package `path`

But below features is not supported.

- Filter  
- Placeholder  
- Function  
- multi properties (ex. $.json1['a','b'])  
- ArrayIndexOperation (ex. $.json1[0,1])  
- ArraySliceOperation (ex. $.json1[1:2] (start:end))

Please review this.