@sonots 2016-08-24
LGTM. Did you check performance although it looks it does not affect worse so much with this implementation?