@kysnm 2016-08-24
> Did you check performance

I'm not check performance. I will check it.