@kysnm 2016-08-24
And I would like to show you the diff between my implementation and Original JsonPath library

https://gist.github.com/kysnm/ba3cc5783a860bdc53ba20b0c2e63cc2

(unescape from jsonpath/internal/Utils https://github.com/jayway/JsonPath/blob/master/json-path/src/main/java/com/jayway/jsonpath/internal/Utils.java#L172-L245)