@kysnm 2016-08-30
I would like to add the example/edgecase.{yml,csv}.
But `*.csv` is ignored.
Why is it?

```diff
diff --git a/.gitignore b/.gitignore
index 734775a..42f9f48 100644
--- a/.gitignore
+++ b/.gitignore
@@ -6,7 +6,6 @@
 /classpath/
 build/
 .idea
-*.csv
 .tags
 .ruby-version
 *.iml
diff --git a/example/edgecase.csv b/example/edgecase.csv
new file mode 100644
index 0000000..a4c6bdb
--- /dev/null
+++ b/example/edgecase.csv
@@ -0,0 +1,4 @@
+json
+"{\"phone']Num\\\"]ber\":\"foo\",\"bar\":\"bar\"}"
+"{\"phone']Num\\\"]ber\":\"foo2\",\"bar\":\"bar2\"}"
+NULL
diff --git a/example/edgecase.yml b/example/edgecase.yml
new file mode 100644
index 0000000..4da9ae2
--- /dev/null
+++ b/example/edgecase.yml
@@ -0,0 +1,20 @@
+in:
+  type: file
+  path_prefix: example/edgecase.csv
+  parser:
+    type: csv
+    charset: UTF-8
+    newline: CRLF
+    null_string: 'NULL'
+    skip_header_lines: 1
+    comment_line_marker: '#'
+    columns:
+      - {name: json, type: json}
+filters:
+  - type: column
+    columns:
+      - {name: json, default: "{\"phone']Num\\\"]ber\":\"FOO\"}"}
+      - {name: "$.json[\"phone']Num\\\"]ber\"]"}
+      - {name: $.json.copy_foo, src: "$.json[\"phone']Num\\\"]ber\"]"}
+out:
+  type: stdout
```