@sonots 2016-08-30
add with git add -f