@kysnm 2016-08-30
I tested the performance, referenced https://github.com/arenahito/embulk-filter-timestamp_hs#performance

I think there is no performance problem.

More derails, Plese see https://gist.github.com/kysnm/62412edb93903bec3ffb033c6bf3fdb0