@kysnm 2016-08-31
Fixed issues. Please review again.