@sonots 2016-09-01
Could you check memory usage too because travis fails with OutOfMemory? Does it really consume lots of memory now?