@kysnm 2016-09-01
I will check the memory usage.
(But it has been successful excect oraclejdk7…)