@kysnm 2016-09-01
### Summary
memory usage looks like as below
(about 500MB up)

master branch

```bash
1317773312  maximum resident set size
```

this branch

```
1883742208  maximum resident set size
```

### Details

```bash
➜  embulk-filter-column git:(master) ./gradlew classpath
:compileJava
warning: [options] bootstrap class path not set in conjunction with -source 1.7
1 warning
:processResources UP-TO-DATE
:classes
:jar
:classpath

BUILD SUCCESSFUL

Total time: 4.662 secs

This build could be faster, please consider using the Gradle Daemon: https://docs.gradle.org/2.10/userguide/gradle_daemon.html
➜  embulk-filter-column git:(master) /usr/bin/time -l ./gradlew test
:compileJava UP-TO-DATE
:processResources UP-TO-DATE
:classes UP-TO-DATE
:compileTestJava
warning: [options] bootstrap class path not set in conjunction with -source 1.7
1 warning
:processTestResources UP-TO-DATE
:testClasses
:test
objc[61462]: Class JavaLaunchHelper is implemented in both /Library/Java/JavaVirtualMachines/jdk1.8.0_25.jdk/Contents/Home/bin/java and /Library/Java/JavaVirtualMachines/jdk1.8.0_25.jdk/Contents/Home/jre/lib/libinstrument.dylib. One of the two will be used. Which one is undefined.

BUILD SUCCESSFUL

Total time: 28.761 secs

This build could be faster, please consider using the Gradle Daemon: https://docs.gradle.org/2.10/userguide/gradle_daemon.html
       29.24 real       104.09 user         3.55 sys
1317773312  maximum resident set size
         0  average shared memory size
         0  average unshared data size
         0  average unshared stack size
    663125  page reclaims
        60  page faults
         0  swaps
        16  block input operations
        50  block output operations
        69  messages sent
        60  messages received
        60  signals received
       154  voluntary context switches
    193965  involuntary context switches
➜  embulk-filter-column git:(master) git checkout support_bracket_notation_for_jsonpath
Switched to branch 'support_bracket_notation_for_jsonpath'
➜  embulk-filter-column git:(support_bracket_notation_for_jsonpath) ./gradlew classpath
:compileJava
warning: [options] bootstrap class path not set in conjunction with -source 1.7
1 warning
:processResources UP-TO-DATE
:classes
:jar
:classpath

BUILD SUCCESSFUL

Total time: 4.626 secs

This build could be faster, please consider using the Gradle Daemon: https://docs.gradle.org/2.10/userguide/gradle_daemon.html
➜  embulk-filter-column git:(support_bracket_notation_for_jsonpath) /usr/bin/time -l ./gradlew test
:compileJava UP-TO-DATE
:processResources UP-TO-DATE
:classes UP-TO-DATE
:compileTestJava
warning: [options] bootstrap class path not set in conjunction with -source 1.7
1 warning
:processTestResources UP-TO-DATE
:testClasses
:test
objc[61523]: Class JavaLaunchHelper is implemented in both /Library/Java/JavaVirtualMachines/jdk1.8.0_25.jdk/Contents/Home/bin/java and /Library/Java/JavaVirtualMachines/jdk1.8.0_25.jdk/Contents/Home/jre/lib/libinstrument.dylib. One of the two will be used. Which one is undefined.

BUILD SUCCESSFUL

Total time: 36.859 secs

This build could be faster, please consider using the Gradle Daemon: https://docs.gradle.org/2.10/userguide/gradle_daemon.html
       37.32 real       124.92 user         4.42 sys
1883742208  maximum resident set size
         0  average shared memory size
         0  average unshared data size
         0  average unshared stack size
    846405  page reclaims
         0  page faults
         0  swaps
         7  block input operations
        70  block output operations
       122  messages sent
       108  messages received
        62  signals received
        85  voluntary context switches
    222937  involuntary context switches
    ```