@kysnm 2016-09-01
a results of in my environment, Both master branch and this branch, memory usage is in about 330~340M. I think there is not much difference between master branch and this branch.