@kysnm 2016-09-03
Changed to determine the quotes using a boolean