@kysnm 2016-09-05
How do you think is this? https://github.com/sonots/embulk-filter-column/pull/14#discussion_r76970908
I would like to do Squash, if there is no problem.