@kysnm 2016-09-06
Fllow the Apache License 2.0 (4.c)
http://www.apache.org/licenses/LICENSE-2.0