@sonots 2016-09-07
As we talked yesterday, let's split jsonpath comiler to another reusable jar library, then use it in this plugin.
So, let me close this PR once. Please revisit.